more diffs

+++ internal/ethapi/api.go
@@ -1761,7 +1761,7 @@ func (s *TransactionAPI) GetTransactionReceipt(ctx context.Context, hash common.
 	if err != nil {
 		return nil, err
 	}
-	if uint64(len(receipts)) <= index {
+	if len(receipts) <= int(index) {
 		return nil, nil
 	}
 	receipt := receipts[index]
